feat(api): add public API endpoints and rate limiting
- Implement public API routes for link creation/deletion and API key generation
- Add express-rate-limit middleware for public API endpoints
- Remove unused backend-worker files and configurations
- Update auth flow to handle API keys and improve session management
- Add validation schemas for public API endpoints
- Include countries list directly in frontend instead of API call
- Improve login/register flows with better error handling
