--- backend/backend-worker/.vscode/settings.json
@@ -1,5 +1,0 @@
-{
-	"files.associations": {
-		"wrangler.json": "jsonc"
-	}
-}
