+++ frontend/src/components/Auth/GuestOrUserRoute.jsx
@@ -1,3 +1,6 @@
+import React from "react";
+import { Navigate } from "react-router-dom";
+
 import { useAuth } from "../../context/Auth";
 import Loading from "../Common/Loading";
